# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/consultations": {
      "get": {
        "operationId": "ConsultationsController_findAll",
        "parameters": [
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expertId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdByExpert",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceCheckinId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsultationEntity"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "List consultations",
        "tags": [
          "consultations"
        ]
      },
      "post": {
        "operationId": "ConsultationsController_createRequest",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsultationRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the expert/employee record is invalid."
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Create a consultation request",
        "tags": [
          "consultations"
        ]
      }
    },
    "/consultations/{id}": {
      "get": {
        "operationId": "ConsultationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Consultation not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Get a consultation by id",
        "tags": [
          "consultations"
        ]
      },
      "patch": {
        "operationId": "ConsultationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConsultationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed for consultation status or session details."
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Consultation not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Update a consultation status or add/edit accepted consultation session details",
        "tags": [
          "consultations"
        ]
      },
      "delete": {
        "operationId": "ConsultationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          },
          "404": {
            "description": "Consultation not found."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Delete a consultation",
        "tags": [
          "consultations"
        ]
      }
    },
    "/consultations/open-latest/search": {
      "get": {
        "operationId": "ConsultationsController_getLatestOpen",
        "parameters": [
          {
            "name": "employeeId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expertId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "companyName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationLookupResultEntity"
                }
              }
            }
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Get the latest open consultation for a given employee and wellness expert pair",
        "tags": [
          "consultations"
        ]
      }
    },
    "/consultations/follow-up": {
      "post": {
        "operationId": "ConsultationsController_createFollowUp",
        "parameters": [
          {
            "name": "role",
            "in": "header",
            "description": "Role used for RBAC. Allowed values: Admin, HR, Employee, Wellness Expert",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFollowUpConsultationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationFollowUpResultEntity"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or the employee/expert record is invalid."
          },
          "403": {
            "description": "Missing or invalid role header."
          }
        },
        "security": [
          {
            "role": []
          }
        ],
        "summary": "Create or reuse an expert follow-up consultation from an employee check-in workflow",
        "tags": [
          "consultations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wellness Platform API",
    "description": "REST APIs for the evaluation backend. Authorization is enforced using the request header role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "role": {
        "type": "apiKey",
        "in": "header",
        "name": "role",
        "description": "Role required for RBAC. Supported values: Admin, HR, Employee, Wellness Expert"
      }
    },
    "schemas": {
...
      "ConsultationEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "consult-1777718400000-a1b2c3d4"
          },
          "employeeId": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "employeeName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "expertId": {
            "type": "string",
            "example": "expert-1777718400000-a1b2c3d4"
          },
          "expertName": {
            "type": "string",
            "example": "Neha Verma"
          },
          "purpose": {
            "type": "string",
            "example": "Need guidance for nutrition planning."
          },
          "category": {
            "type": "string",
            "example": "Nutritionist"
          },
          "requestedOn": {
            "type": "string",
            "example": "02 May 2026 @ 6:15 pm"
          },
          "status": {
            "type": "string",
            "example": "requested"
          },
          "rejectionReason": {
            "type": "string",
            "example": ""
          },
          "sessionTitle": {
            "type": "string",
            "example": "Nutrition Consultation with Aarav Sharma"
          },
          "sessionDate": {
            "type": "string",
            "example": "2026-05-03"
          },
          "sessionTime": {
            "type": "string",
            "example": "18:30"
          },
          "sessionDuration": {
            "type": "string",
            "example": "45 min"
          },
          "sessionMeetingLink": {
            "type": "string",
            "example": "https://meet.example.com/session"
          },
          "sessionCreatedAt": {
            "type": "string",
            "example": "2026-05-02T13:00:00.000Z"
          },
          "createdByExpert": {
            "type": "boolean",
            "example": false
          },
          "sourceCheckinId": {
            "type": "string",
            "example": "checkin-1777718400000-a1b2c3d4"
          },
          "followUpPriority": {
            "type": "string",
            "example": "high"
          },
          "companyId": {
            "type": "string",
            "example": "CMP-1001"
          },
          "companyName": {
            "type": "string",
            "example": "Stack Builders Pvt Ltd"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-02T12:30:00.000Z"
          }
        },
        "required": [
          "id",
          "employeeId",
          "employeeName",
          "expertId",
          "expertName",
          "purpose",
          "category",
          "requestedOn",
          "status",
          "rejectionReason",
          "sessionTitle",
          "sessionDate",
          "sessionTime",
          "sessionDuration",
          "sessionMeetingLink",
          "sessionCreatedAt",
          "createdByExpert",
          "sourceCheckinId",
          "followUpPriority",
          "companyId",
          "companyName",
          "createdAt"
        ]
      },
      "ConsultationLookupResultEntity": {
        "type": "object",
        "properties": {
          "consultation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/ConsultationEntity"
              }
            ]
          }
        }
      },
      "CreateConsultationRequestDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "employeeName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "expertId": {
            "type": "string",
            "example": "expert-1777718400000-a1b2c3d4"
          },
          "expertName": {
            "type": "string",
            "example": "Neha Verma"
          },
          "purpose": {
            "type": "string",
            "example": "Need guidance for nutrition planning."
          }
        },
        "required": [
          "employeeId",
          "expertId",
          "purpose"
        ]
      },
      "CreateFollowUpConsultationDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "example": "employee-1777718400000-a1b2c3d4"
          },
          "employeeName": {
            "type": "string",
            "example": "Aarav Sharma"
          },
          "expertId": {
            "type": "string",
            "example": "expert-1777718400000-a1b2c3d4"
          },
          "expertName": {
            "type": "string",
            "example": "Neha Verma"
          },
          "purpose": {
            "type": "string",
            "example": "High nutrition follow-up recommended based on the latest employee update."
          },
          "sourceCheckinId": {
            "type": "string",
            "example": "checkin-1777718400000-a1b2c3d4"
          },
          "followUpPriority": {
            "type": "string",
            "example": "high"
          }
        },
        "required": [
          "employeeId",
          "expertId"
        ]
      },
      "ConsultationFollowUpResultEntity": {
        "type": "object",
        "properties": {
          "consultation": {
            "$ref": "#/components/schemas/ConsultationEntity"
          },
          "reused": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "consultation",
          "reused"
        ]
      },
      "UpdateConsultationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "requested"
          },
          "rejectionReason": {
            "type": "string",
            "example": "Consultation request could not be scheduled."
          },
          "sessionTitle": {
            "type": "string",
            "example": "Nutrition Consultation with Aarav Sharma"
          },
          "sessionDate": {
            "type": "string",
            "example": "2026-05-03"
          },
          "sessionTime": {
            "type": "string",
            "example": "18:30"
          },
          "sessionDuration": {
            "type": "string",
            "example": "45 min"
          },
          "sessionMeetingLink": {
            "type": "string",
            "example": "https://meet.example.com/session"
          },
          "sessionCreatedAt": {
            "type": "string",
            "example": "2026-05-02T13:00:00.000Z"
          },
          "createdByExpert": {
            "type": "boolean",
            "example": true
          },
          "sourceCheckinId": {
            "type": "string",
            "example": "checkin-1777718400000-a1b2c3d4"
          },
          "followUpPriority": {
            "type": "string",
            "example": "high"
          },
          "purpose": {
            "type": "string",
            "example": "Need guidance for nutrition planning."
          }
        }
      },
...
    }
  }
}